FIX - blank datastore name when default datastore is used
if GOVC_DATASTORE is not set then the default datastore is derived using
DefaultDatastore.
The ISO is written to the datastore location calculated, but the CDROM
was configured to use a blank datastore.

Error message:
FATA[0042] Failed to create master: clone: failed create waiter: Invalid
configuration for device '3'.

since for types.VirtualDeviceFileBackingInfo the FileName is set to a
blank Datastore, e.g.:

FileName:                 "[] cloud-init/tka-test-kubeadm-configs-m-161-143555.iso",